Move all internal packages in the same spot, move main to root of repo
