+++ .gitignore
@@ -15,3 +15,5 @@
 .DS_Store
 
 build/
+
+dist/
